refactor: 페이지에 맞게 컴포넌트 정리
